Struct atmega48p_hal::usart::Usart0

pub struct Usart0<CLOCK, RX_MODE> where
    CLOCK: Clock,
    RX_MODE: InputMode,  { /* fields omitted */ }

Serial interface based on ATmega48P's USART0 peripheral

Maximum baudrate seems to be 57600

Implementations

impl<CLOCK, RX_MODE> Usart0<CLOCK, RX_MODE> where
CLOCK: Clock,
RX_MODE: InputMode,

pub fn new(
p: USART0,
rx: PD0<Input<RX_MODE>>,
tx: PD1<Output>,
baud: Baudrate<CLOCK>
) -> Usart0<CLOCK, RX_MODE>

Initialize the USART peripheral

Please note that not all baudrates will produce a good signal and setting it too high might make data sent completely unreadable for the other side.

pub fn interrupt_rxc(&mut self, state: bool)

Enable/disable "RX Complete" interrupt

When this interrupt triggers, new data is available to be read from the data-register. The corresponding ISR is USARTi_RX (where i is this peripheral's number). For example, for USART1 on ATmega32U4:

#[avr_device::interrupt(atmega32u4)]
fn USART1_RX() {
    // ...
}

pub fn interrupt_udre(&mut self, state: bool)

Enable/disable "USART Data-Register Empty" interrupt

This interrupt signals that new data can be written to the data-register. The corresponding ISR is USARTi_UDRE (where i is this peripheral's number). For example, for USART1 on ATmega32U4:

#[avr_device::interrupt(atmega32u4)]
fn USART1_UDRE() {
    // ...
}

pub fn split(self) -> (ReadUsart0<CLOCK, RX_MODE>, WriteUsart0<CLOCK>)

Helper method for splitting this read/write object into two halves.

The two halves returned implement the Read and Write traits, respectively.